aichan/pyproject.toml
[project]
name = "aichan"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "agent",
    "core",
    "fastapi>=0.118.0",
    "httpx>=0.28.0",
    "httpx-sse>=0.4.0",
    "mcp_hub",
    "langchain-core>=1.2.20",
    "langchain-openai>=1.1.11",
    "memory",
    "uvicorn>=0.37.0",
]

[tool.uv.workspace]
members = ["agent", "core", "memory", "mcp_hub"]

[tool.uv.sources]
agent = { workspace = true }
core = { workspace = true }
memory = { workspace = true }
mcp_hub = { workspace = true }

